Move meta checks and build triggering
